requirements.txt
PyTelegramBotAPI>=4.6.0
pyyaml>=6.0
sqlmodel==0.0.6
requests>=2.28.1